Bring back pre-dblp changes into main dblp branch

+++ data/types/food/carbohydrate.go
@@ -6,14 +6,14 @@ import (
 )
 
 const (
-	CarbohydrateDietaryFiberGramsMaximum = 1000
-	CarbohydrateDietaryFiberGramsMinimum = 0
-	CarbohydrateNetGramsMaximum          = 1000
-	CarbohydrateNetGramsMinimum          = 0
-	CarbohydrateSugarsGramsMaximum       = 1000
-	CarbohydrateSugarsGramsMinimum       = 0
-	CarbohydrateTotalGramsMaximum        = 1000
-	CarbohydrateTotalGramsMinimum        = 0
+	CarbohydrateDietaryFiberGramsMaximum = 1000.0
+	CarbohydrateDietaryFiberGramsMinimum = 0.0
+	CarbohydrateNetGramsMaximum          = 1000.0
+	CarbohydrateNetGramsMinimum          = 0.0
+	CarbohydrateSugarsGramsMaximum       = 1000.0
+	CarbohydrateSugarsGramsMinimum       = 0.0
+	CarbohydrateTotalGramsMaximum        = 1000.0
+	CarbohydrateTotalGramsMinimum        = 0.0
 	CarbohydrateUnitsGrams               = "grams"
 )
 
@@ -24,11 +24,11 @@ func CarbohydrateUnits() []string {
 }
 
 type Carbohydrate struct {
-	DietaryFiber *int    `json:"dietaryFiber,omitempty" bson:"dietaryFiber,omitempty"`
-	Net          *int    `json:"net,omitempty" bson:"net,omitempty"`
-	Sugars       *int    `json:"sugars,omitempty" bson:"sugars,omitempty"`
-	Total        *int    `json:"total,omitempty" bson:"total,omitempty"`
-	Units        *string `json:"units,omitempty" bson:"units,omitempty"`
+	DietaryFiber *float64 `json:"dietaryFiber,omitempty" bson:"dietaryFiber,omitempty"`
+	Net          *float64 `json:"net,omitempty" bson:"net,omitempty"`
+	Sugars       *float64 `json:"sugars,omitempty" bson:"sugars,omitempty"`
+	Total        *float64 `json:"total,omitempty" bson:"total,omitempty"`
+	Units        *string  `json:"units,omitempty" bson:"units,omitempty"`
 }
 
 func ParseCarbohydrate(parser data.ObjectParser) *Carbohydrate {
@@ -46,18 +46,18 @@ func NewCarbohydrate() *Carbohydrate {
 }
 
 func (c *Carbohydrate) Parse(parser data.ObjectParser) {
-	c.DietaryFiber = parser.ParseInteger("dietaryFiber")
-	c.Net = parser.ParseInteger("net")
-	c.Sugars = parser.ParseInteger("sugars")
-	c.Total = parser.ParseInteger("total")
+	c.DietaryFiber = parser.ParseFloat("dietaryFiber")
+	c.Net = parser.ParseFloat("net")
+	c.Sugars = parser.ParseFloat("sugars")
+	c.Total = parser.ParseFloat("total")
 	c.Units = parser.ParseString("units")
 }
 
 func (c *Carbohydrate) Validate(validator structure.Validator) {
-	validator.Int("dietaryFiber", c.DietaryFiber).InRange(CarbohydrateDietaryFiberGramsMinimum, CarbohydrateDietaryFiberGramsMaximum)
-	validator.Int("net", c.Net).Exists().InRange(CarbohydrateNetGramsMinimum, CarbohydrateNetGramsMaximum)
-	validator.Int("sugars", c.Sugars).InRange(CarbohydrateSugarsGramsMinimum, CarbohydrateSugarsGramsMaximum)
-	validator.Int("total", c.Total).InRange(CarbohydrateTotalGramsMinimum, CarbohydrateTotalGramsMaximum)
+	validator.Float64("dietaryFiber", c.DietaryFiber).InRange(CarbohydrateDietaryFiberGramsMinimum, CarbohydrateDietaryFiberGramsMaximum)
+	validator.Float64("net", c.Net).Exists().InRange(CarbohydrateNetGramsMinimum, CarbohydrateNetGramsMaximum)
+	validator.Float64("sugars", c.Sugars).InRange(CarbohydrateSugarsGramsMinimum, CarbohydrateSugarsGramsMaximum)
+	validator.Float64("total", c.Total).InRange(CarbohydrateTotalGramsMinimum, CarbohydrateTotalGramsMaximum)
 	validator.String("units", c.Units).Exists().OneOf(CarbohydrateUnits()...)
 }
